Updates dashboard URL scheme
