+++ app/brochure/views/partials/header.html
@@ -22,7 +22,7 @@
     {{/selected.log-in}}
 
     {{#user}}
-    <a href="/">Your blogs</a>
+    <a href="/dashboard">Your blogs</a>
     {{/user}}
 
     {{^user}}
