+++ app/dashboard/render.js
@@ -1,4 +1,6 @@
 module.exports = function (req, res, next) {
+  res.locals.partials = res.locals.partials || {};
+
   var _render = res.render;
   var wrapper;
   res._render = _render;
