+++ app/dashboard/routes/account/index.js
@@ -73,7 +73,6 @@ Account.use("/:section/:subsection", function (req, res, next) {
 Account.use("/password", require("./password"));
 Account.use("/export", require("./export"));
 Account.use("/email", require("./email"));
-Account.use("/switch-blog", require("./switch-blog"));
 Account.use("/create-blog", require("./create-blog"));
 Account.use("/subscription", require("./subscription"));
 Account.use("/pay-subscription", require("./pay-subscription"));
