+++ app/dashboard/routes/folder/index.js
@@ -7,11 +7,14 @@ var trace = require("helper/trace");
 
 router
   .use(function (req, res, next) {
-    res.locals.partials.entry = "folder/entry";
-    res.locals.partials.stat = "folder/stat";
-    res.locals.partials.file = "folder/file";
-    res.locals.partials.directory = "folder/directory";
-    res.locals.partials.folder = "folder/wrapper";
+    res.locals.partials = {
+      ...res.locals.partials,
+      entry: "folder/entry",
+      stat: "folder/stat",
+      file: "folder/file",
+      directory: "folder/directory",
+      folder: "folder/wrapper",
+    };
     next();
   })
   .use(trace("determining path"))
@@ -21,17 +24,6 @@ router
   .use(trace("renderFolder"))
   .use(renderFolder)
   .use(trace("renderFile"))
-  .use(renderFile)
-  .use(function (err, req, res, next) {
-    console.log(err);
-    // suppress errors
-    next();
-  });
-
-router.post("/path", function (req, res) {
-  req.session[req.blog.id] = req.session[req.blog.id] || {};
-  req.session[req.blog.id].path = req.body.path || "/";
-  return res.redirect(req.body.redirect || "/");
-});
+  .use(renderFile);
 
 module.exports = router;
