--- app/dashboard/routes/tools/index.js
@@ -1,3 +1,0 @@
-module.exports = function (server) {
-  require("./verify-domain")(server);
-};
