+++ app/dashboard/views/account/pay-subscription.html
@@ -11,29 +11,7 @@ a:hover {text-decoration: underline;}
 
 
 <p style="margin-top:0;padding:11px 20px 0;margin-bottom:0">Your subscription fee is overdue. Please pay {{amountDue}} to restart your subscription. You can also permanently <a href="/account/subscription/delete">delete your account</a>. If you have any questions please <a href="/support">contact support</a>.</p>
-<!-- 
-<div class="center" style="position:relative;top:-3px;padding:0 20px;max-width:75%">
-<p style="margin:6px 0 0;padding:0 0 6px;border-bottom:1px solid rgba(0,0,0,0.05);color:rgb(158, 154, 152);">Your blogs</p>
-  {{#blogs}}
 
-  <div style="display:flex;border:none;align-items:center;padding:0;border-bottom:1px solid rgba(0,0,0,0.05)" class="line">
-
-    <form action="/account/switch-blog" style="display:flex;flex-grow:1;" method="post">
-      <input type="hidden" name="_csrf" value="{{csrftoken}}">
-      <input type="hidden" name="to" value="{{id}}">
-      <button type="submit" style="width:100%;padding:6px 11px 6px 0;display:flex;border:none;margin:0;background:none">
-        <span style="display:flex;flex-grow:1;align-items:center"><span class="photo-holder" style="margin-top:0"> {{#avatar}}<img style="vertical-align: top;" src="{{avatar}}?handle={{handle}}">{{/avatar}}
-              {{^avatar}}<span class="file" style="font-size:0.8em;opacity:0.5;margin:-0.3em 0 0 0"></span>{{/avatar}}</span><span class="link">{{pretty.label}}</span></span>
-      </button>
-    </form>
-
-    <a class="right link" style="display:flex;background:none;padding:0" href="/account/delete/blog/{{handle}}">Delete </a>
-
-  </div>
-
-  {{/blogs}}
-  <div class="clear"></div>
-</div> -->
 
 <form method="POST" class="account" id="payment-form">
