+++ app/dashboard/views/folder/wrapper.html
@@ -3,7 +3,7 @@
   <div style="overflow:hidden;display:flex;flex-direction:column; box-shadow: rgba(15,15,15,.1) 0 0 0 1px, rgba(15,15,15,.1) 0 2px 4px;
     background: #fff;
     border-radius: 6px;">
-{{#root}}<a class="line" href="/settings/client/{{client.name}}">
+{{#root}}<a class="line" href="{{{base}}}/client/{{client.name}}">
 <span class="label">Folder</span>
 {{^blog.client}}
       <span class="link">Set up a client to sync your folder</span>
