+++ app/dashboard/views/settings/domain.html
@@ -1,6 +1,5 @@
 <p>Your site comes with a Blot subdomain. You can also <a href="/how/configure/domain">use your own domain</a>. Please don't hesitate to <a href="/support">contact support</a> if you have any questions.</p>
 <hr>
-<input type="hidden" name="redirect" value="/settings/domain">
   <label>Your domain</label>
   <div class="messageWrapper">
     <input type="text" id="customDomain" name="domain" 
@@ -28,7 +27,7 @@
   <script src="/scripts/jquery.js?{{cacheID}}"></script>
   <script type="text/javascript">
     if ($('#customDomain').val()) {
-     $.get('/verify-domain/' + $('#customDomain').val(), function(data) {
+     $.get('{{{base}}}/verify-domain/' + $('#customDomain').val(), function(data) {
        if (data && data === 'true' || data === true) {
          $('#inputMessage').html('').addClass('success');
        } else {
