+++ app/dashboard/views/settings/index.html
@@ -9,7 +9,7 @@
 <div>
 
   <form method='post' action='?_csrf={{csrftoken}}' enctype='multipart/form-data'>
-    <a class="line" style="border-top:1px solid rgba(0,0,0,0.05)" href="/settings/title">
+    <a class="line" style="border-top:1px solid rgba(0,0,0,0.05)" href="{{{base}}}/title">
       <span class="label">Title</span>
       <span class="center">
         {{blog.title}}
@@ -22,13 +22,13 @@
       {{/blog.title}}
       
     </a>
-    <a class="line" href="/settings/domain">
+    <a class="line" href="{{{base}}}/domain">
       <span class="label">Domain</span>
       <span class="center">{{blog.pretty.url}}</span>
       <span class="right">Edit</span>
     </a>
 
-    <a class="line" href="/settings/template">
+    <a class="line" href="{{{base}}}/template">
       <span class="label">Template</span>
       <span class="center">
         {{#template.name}}
@@ -42,7 +42,7 @@
       <span class="right">Edit</span>
     </a>
 
-    <a class="line" href="/settings/photo">
+    <a class="line" href="{{{base}}}/photo">
       <span class="label" style="float:left">Photo</span>  
       {{^blog.avatar}}
       <span class="link">Choose a photo</span>
@@ -58,7 +58,7 @@
     </section>
   </a>
 
-  <a class="line" href="/settings/links">
+  <a class="line" href="{{{base}}}/links">
     <span class="label">Links</span>
     <span class="center" style="overflow:hidden;position:relative;left:-2px">
       <div style="height:1.5em;width:1000px;">
@@ -75,7 +75,7 @@
 
   </a>
 
-  <a class="line" href="/settings/services">
+  <a class="line" href="{{{base}}}/services">
     <span class="label">
       Services
     </span>
@@ -92,7 +92,7 @@
     </span>
   </a>
   <!-- 
-    <a class="line" href="/settings/photo">
+    <a class="line" href="{{{base}}}/photo">
       <span class="label">Photo</span>
       <span class="center">
 
@@ -112,7 +112,7 @@
     </a> -->
 
     <!-- 
-    <a class="line" href="/settings/urls" >
+    <a class="line" href="{{{base}}}/urls" >
       <span class="label">URLs</span>
       <span class="center">{{#permalinkFormats}}{{#checked}}{{example}}{{/checked}}{{/permalinkFormats}}</span>
       <span class="right">Change</span>
