+++ app/dashboard/views/settings/links.html
@@ -1,9 +1,8 @@
-<p>The following links appear on your site. You can add new links to external sites. <a href="/how/pages">Pages</a> in your folder are added to these links automatically.</p>
+<p>The following links appear on your site. You can add new links to external sites. <a href="/how/pages">Pages</a> in your folder are added to these links automatically. Click-and-drag the handle on the links to re-order them.</p>
 <hr>
 
 <!-- Neccessary to produce an empty menu -->
 <input type="hidden" name="hasMenu" value="true">
-<input type="hidden" name="redirect" value="/settings/links">
 
 <section id="menu" class="fullRow sortable"  style="max-width:100%">
