+++ app/dashboard/views/settings/photo.html
@@ -1,6 +1,5 @@
 <p>Your photo is used on your site's template and to generate a favicon for your site.</p>
 
-<input type="hidden" name="redirect" value="/settings/photo">
 
   {{#blog.avatar}} 
 <div class="line" style="padding-left: 0">
