+++ app/dashboard/views/settings/services.html
@@ -1,6 +1,5 @@
 <style type="text/css">form {max-width:none}</style>
 <div style="margin:-11px -26px">
-  <input type="hidden" name="redirect" value="/settings/services">
 
   <a class="line" href="/settings/services/date">
     <span class="label">Date and time</span>
