+++ app/dashboard/views/settings/subpage.html
@@ -2,18 +2,18 @@
 form {border-top: 1px solid rgba(0,0,0,0.05);max-width:700px;}
 form label {padding-top: 0} hr {margin: 20px 0} 
 </style>
-<form method='post' {{^setup}}style="padding: 11px 26px"{{/setup}} {{#setup}}style="padding:20px 0"{{/setup}} action='/settings' enctype='multipart/form-data'>
+<form method='post' {{^setup}}style="padding: 11px 26px"{{/setup}} {{#setup}}style="padding:20px 0"{{/setup}} action='{{{base}}}' enctype='multipart/form-data'>
 
 
   <input type="hidden" name="_csrf" value="{{csrftoken}}">
   <input type="hidden" name="setup" value="{{setup}}">
-  <input type="hidden" name="redirect" value="/settings/{{subpage_slug}}">
+  <input type="hidden" name="redirect" value="{{{base}}}/{{subpage}}">
     
   {{> subpage}}
 
   <div class="buttons">
     <button type="submit"onclick="this.form.submit();this.disabled=true">Save changes</button>
-    <a onclick="this.classList.add('disabled')" href="/settings">Cancel</a>
+    <a onclick="this.classList.add('disabled')" href="{{{base}}}">Cancel</a>
   </div>
   <br>
 </form>
