+++ app/dashboard/views/settings/title.html
@@ -22,5 +22,5 @@
 </h2>
 <hr /> -->
 <label for="title" style="margin-top: 1rem">Title</label>
-<input type="hidden" name="redirect" value="/settings/title" />
+<input type="hidden" name="redirect" value="{{{base}}}/title" />
 <input type="text" name="title" value="{{blog.title}}" autofocus />
